requirements/base.txt
fastapi==0.65.2
uvicorn[standard]==0.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
harperdb==0.4.2